Maximo_Weather_Scheduler-main/requirements.txt
pandas
python-dotenv
requests
openpyxl
flask
flask-cors
wget
ibm_watsonx_ai
ibm_cloud_sdk_core
pyarrow
flask_restx
numpy>=1.21

docling
langchain
langchain-core
ibm-watsonx-ai

duckduckgo-search

beeai-framework==0.1.15

